actualización id_jugador no requerido
